fix: cap hex object IDs to 8 chars by removing 64-bit sign-extension
ObjectIdCompat.GetRawId pre-6000.4 used (ulong)(long) cast which
sign-extended negative 32-bit instance IDs to 64-bit, producing
16-char hex refs like $FFFFFFFFFFFFCFC7. Changed to (uint) cast —
negative IDs now produce max 8-char hex ($FFFFCFC7). Resolution
is lossless since ResolveObject already truncates back to int.

+++ unity-plugin/Editor/ObjectIdCompat.cs
@@ -26,7 +26,7 @@ namespace UnityMCP.Editor
                && property.objectReferenceEntityIdValue != EntityId.None;
 #else
         internal static ulong GetRawId(Object obj)
-            => obj == null ? 0UL : unchecked((ulong)(long)obj.GetInstanceID());
+            => obj == null ? 0UL : unchecked((uint)obj.GetInstanceID());
 
         internal static Object ResolveObject(ulong rawId)
             => rawId == 0UL ? null
